@@ -1,13 +1,14 @@
-﻿using Amazon;
-using Amazon.DynamoDBv2;
+﻿using Amazon.DynamoDBv2;
 using Amazon.DynamoDBv2.Model;
-using Amazon.Runtime;
 using Amazon.S3;
 using Amazon.S3.Model;
 using Amazon.SimpleNotificationService;
 using Amazon.SimpleNotificationService.Model;
 using Config.Interfaces;
 using Core.Exceptions;
+using MimeKit.Text;
+using MimeKit;
+using NuevaLuz.Fonoteca.Helper;
 using NuevaLuz.Fonoteca.Models;
 using NuevaLuz.Fonoteca.Services.Fonoteca.Interfaces;
 using System;
@@ -15,8 +16,6 @@ using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Globalization;
 using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
@@ -27,35 +26,38 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
         string _connectionString { get; }
         ISettings _settings { get; }
 
+        private AmazonDynamoDBClient _clientDynamoDB;
+        private AmazonSimpleNotificationServiceClient _clientSNS;
+
         public FonotecaService(ISettings settings)
         {
             _connectionString = $"Server={settings.Server},{settings.Port};Database={settings.Database};User Id={settings.User};Password={settings.Password};";
             _settings = settings;
+            _clientDynamoDB = new AmazonDynamoDBClient();
+            _clientSNS = new AmazonSimpleNotificationServiceClient();
         }
 
-        private async Task CheckCredentials(int user, string password)
+        private async Task<bool> CheckCredentials(int user, string password)
         {
             using SqlConnection connection = new SqlConnection(
    _connectionString);
-            SqlCommand commandCount = new SqlCommand($@"SELECT contrasena FROM US_usuarios WHERE id=${user} AND activoweb=1 AND activo=1",
+            SqlCommand commandCount = new SqlCommand($@"SELECT contrasena_encrip, contrasena_act FROM US_usuarios WHERE id=${user} AND activoweb=1 AND activo=1",
     connection);
 
             connection.Open();
 
             using SqlDataReader reader = await commandCount.ExecuteReaderAsync();
-            if (!(reader.Read() && reader[0].ToString().Trim() == password))
+            if (!(reader.Read() && reader[0].ToString().Trim() == HashHelper.GenerateHash(password)))
             {
                 throw new Exception("Usuario o contraseña incorrectos");
             }
+            return bool.Parse(reader[1].ToString());
         }
 
         public async Task CheckNotificationsAccess(string user, string password)
         {
-            string hashPass = Convert.ToBase64String(
-                new MD5CryptoServiceProvider().ComputeHash(
-                    new UnicodeEncoding().GetBytes(password)
-                )
-            );
+            // Compute the hash of the password using MD5
+            string hashPass = HashHelper.GenerateHash(password);
 
             using SqlConnection connection = new SqlConnection(
    _connectionString);
@@ -82,12 +84,6 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
         {
             if (string.IsNullOrEmpty(session)) throw new Exception("Sesión no válida");
 
-            // Insert in dynamo db
-            AmazonDynamoDBClient clientDynamoDB = new AmazonDynamoDBClient(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
-
             var request = new QueryRequest
             {
                 TableName = _settings.SessionTableName,
@@ -97,7 +93,7 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
                 }
             };
 
-            var response = await clientDynamoDB.QueryAsync(request);
+            var response = await _clientDynamoDB.QueryAsync(request);
 
             var sessionInfo = response.Items.FirstOrDefault();
 
@@ -116,17 +112,11 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
             }
         }
 
-        public async Task<string> Login(int user, string password)
+        public async Task<LoginResult> Login(int user, string password)
         {
-            await CheckCredentials(user, password);
+            var mustBeChanged = await CheckCredentials(user, password);
 
             string result = Guid.NewGuid().ToString();
-
-            // Insert in dynamo db
-            AmazonDynamoDBClient clientDynamoDB = new AmazonDynamoDBClient(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
 
             // Define item attributes
             Dictionary<string, AttributeValue> attributes = new Dictionary<string, AttributeValue>
@@ -151,18 +141,13 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
                 Item = attributes
             };
 
-            await clientDynamoDB.PutItemAsync(request);
+            await _clientDynamoDB.PutItemAsync(request);
 
-            return result;
+            return new LoginResult { Session = result, MustBeChanged = mustBeChanged };
         }
 
         public async Task ChangePassword(string session, string password)
         {
-            AmazonDynamoDBClient clientDynamoDB = new AmazonDynamoDBClient(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
-
             var request = new QueryRequest
             {
                 TableName = _settings.SessionTableName,
@@ -172,7 +157,7 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
                 }
             };
 
-            var response = await clientDynamoDB.QueryAsync(request);
+            var response = await _clientDynamoDB.QueryAsync(request);
 
             var sessionInfo = response.Items.FirstOrDefault();
 
@@ -182,7 +167,7 @@ namespace NuevaLuz.Fonoteca.Services.Fonoteca.Implementations
 
                 using SqlConnection connection = new SqlConnection(
                    _connectionString);
-                SqlCommand commandUpdate = new SqlCommand($@"UPDATE US_usuarios SET contrasena='{password}' WHERE id={currentUser}",
+                SqlCommand commandUpdate = new SqlCommand($@"UPDATE US_usuarios SET contrasena_encrip='{HashHelper.GenerateHash(password)}', contrasena_act = 0 WHERE id={currentUser}",
 connection);
 
                 connection.Open();
@@ -198,6 +183,112 @@ connection);
             {
                 throw new AuthenticationException(ExceptionCodes.IDENTITY_AUTHORIZATION_SECURITY_EXCEPTION, "Sesión inválida");
             }
+        }
+
+        public async Task ForgotPassword(string email)
+        {
+            using (SqlConnection connection = new SqlConnection(
+   _connectionString))
+            {
+                SqlCommand commandCount = new SqlCommand(@$"SELECT id total FROM US_usuarios WHERE US_usuarios.email = '{email}'",
+                              connection);
+
+                connection.Open();
+
+                var userId = 0;
+                using (SqlDataReader reader = await commandCount.ExecuteReaderAsync())
+                {
+                    while (reader.Read())
+                    {
+                        userId = Convert.ToInt32(reader[0]);
+                        break;
+                    }
+                }
+
+                string cleanPassword = "";
+                if (userId > 0)
+                {
+                    // Create a new password 6 chars longs
+                    cleanPassword = HashHelper.GenerateRandomPassword();
+                    var newPassword = HashHelper.GenerateHash(cleanPassword);
+
+                    SqlCommand commandUpdate = new SqlCommand($@"UPDATE US_usuarios SET contrasena_encrip='{newPassword}', contrasena_act = 1 WHERE id={userId}",
+    connection);                    
+
+                    await commandUpdate.ExecuteNonQueryAsync();
+                }
+
+                // Send email with the new password
+                var mimeMessage = new MimeMessage();                
+                mimeMessage.From.Add(new MailboxAddress(_settings.AwsSmtpFromName, _settings.AwsSmtpFrom));
+                mimeMessage.To.Add(new MailboxAddress(email, email));
+                mimeMessage.Subject = "Nueva contraseña de acceso a Fonoteca Nueva Luz";
+                mimeMessage.Body = new TextPart(TextFormat.Plain)
+                {
+                    Text = $@"Estimado usuario de fonoteca, 
+
+Se han generado una nuevas credenciales de acceso
+Usario: {userId}
+Contraseña: {cleanPassword}
+
+En el nuevo inicio de sesión se le pedirá que cambie la contraseña por una de su elección.
+
+Atentamente,
+Nueva Luz.
+"
+                };
+                await SendEmailHelper.SendEmail(_settings, mimeMessage);
+            }
+        }
+
+        public async Task<TitleResult> GetRecentBooks(int index, int count)
+        {
+            TitleResult result = new TitleResult
+            {
+                Titles = new List<TitleModel>()
+            };
+
+            using (SqlConnection connection = new SqlConnection(
+               _connectionString))
+            {
+                SqlCommand commandCount = new SqlCommand(@"SELECT count(*) total FROM LH_audioteca, LH_formatosdisponibles
+                          WHERE LH_audioteca.id = LH_formatosdisponibles.id_audioteca AND 
+                          LH_formatosdisponibles.id_formato = 4 AND LH_formatosdisponibles.activo = 'True' AND LH_audioteca.activo = 'True'",
+                              connection);
+
+                connection.Open();
+
+                using (SqlDataReader reader = await commandCount.ExecuteReaderAsync())
+                {
+                    while (reader.Read())
+                    {
+                        result.Total = Convert.ToInt32(reader[0]);
+                    }
+                }
+
+                SqlCommand commandTitles = new SqlCommand($@"SELECT * FROM (
+                            SELECT ROW_NUMBER() OVER(ORDER BY LH_audioteca.f_alta DESC) AS idx, 
+                                   LH_audioteca.numero 'id', LH_audioteca.titulo, LH_audioteca.id_autor 
+                               FROM LH_audioteca, LH_formatosdisponibles 
+                               WHERE LH_audioteca.id = LH_formatosdisponibles.id_audioteca AND LH_formatosdisponibles.id_formato = 4 AND LH_formatosdisponibles.activo = 'True' AND LH_audioteca.activo = 'True'
+                         ) AS tbl WHERE idx BETWEEN ${index} AND ${index + count - 1}",
+                         connection);
+
+                using (SqlDataReader reader = await commandTitles.ExecuteReaderAsync())
+                {
+                    while (reader.Read())
+                    {
+                        result.Titles.Add(new TitleModel
+                        {
+                            Id = reader[1].ToString().Trim(),
+                            Title = reader[2].ToString().Trim(),
+                            AuthorId = reader[3].ToString().Trim()
+                        });
+                    }
+                }
+            }
+
+            return result;
         }
 
         public async Task<TitleResult> GetBooksByTitle(int index, int count)
@@ -362,27 +453,15 @@ connection);
             return result;
         }
 
-        public async Task<AudioBookLinkResult> GetAudioBookLink(string session, string id)
+        public async Task IncreaseTitleDownloadCounter(string session, string id)
         {
-            IAmazonS3 clientS3 = new AmazonS3Client(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
-
-            var link = clientS3.GetPreSignedURL(new GetPreSignedUrlRequest
-            {
-                BucketName = _settings.AWSBucket,
-                Expires = DateTime.Now.AddSeconds(_settings.AWSLinkExpireInSecs),
-                Key = $"{id.PadLeft(4, '0')}.zip"
-            });
-
             // Update Database counter so we have to get user from dynamod db (from session)
             var currentUser = await GetUserFromSession(session);
 
             if (!string.IsNullOrEmpty(currentUser))
             {
-                using SqlConnection connection = new SqlConnection(
-                   _connectionString);
+                using SqlConnection connection = new SqlConnection(_connectionString);
+
                 SqlCommand commandInsert = new SqlCommand($@"INSERT INTO LH_historico (id_usuario, id_audioteca, id_formato, id_estado,
                     f_mibiblioteca, f_pendiente, f_envio, f_devolucion, regalo, gestor_mibiblioteca, gestor_pendiente, gestor_envio,
                     gestor_devolucion, web) VALUES (${currentUser}, ${id}, 4, 5, GETDATE(), GETDATE(), GETDATE(), GETDATE(), 'True', 'MOVIL', 'MOVIL', 'MOVIL', 'MOVIL', 'True')",
@@ -392,6 +471,18 @@ connection);
 
                 await commandInsert.ExecuteNonQueryAsync();
             }
+        }
+
+        public AudioBookLinkResult GetAudioBookLink(string session, string id)
+        {
+            IAmazonS3 clientS3 = new AmazonS3Client();
+
+            var link = clientS3.GetPreSignedURL(new GetPreSignedUrlRequest
+            {
+                BucketName = _settings.AWSBucket,
+                Expires = DateTime.Now.AddSeconds(_settings.AWSLinkExpireInSecs),
+                Key = $"{id.PadLeft(4, '0')}.zip"
+            });
 
             return new AudioBookLinkResult
             {
@@ -401,12 +492,6 @@ connection);
 
         private async Task<string> GetUserFromSession(string session)
         {
-            // Update Database counter so we have to get user from dynamod db (from session)
-            AmazonDynamoDBClient clientDynamoDB = new AmazonDynamoDBClient(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
-
             var request = new QueryRequest
             {
                 TableName = _settings.SessionTableName,
@@ -416,7 +501,7 @@ connection);
                 }
             };
 
-            var response = await clientDynamoDB.QueryAsync(request);
+            var response = await _clientDynamoDB.QueryAsync(request);
 
             var sessionInfo = response.Items.FirstOrDefault();
 
@@ -436,7 +521,7 @@ connection);
                _connectionString))
             {
                 SqlCommand commandDeatils = new SqlCommand($@"SELECT LHA.numero 'id', LHA.titulo, LHA.comentario, LHA.id_autor, LHA.horas, LHA.minutos,
-                       SIA.nombre 'autor', SIE.nombre 'editorial'
+                       SIA.nombre 'autor', SIE.nombre 'editorial', LHA.f_alta
                        FROM LH_audioteca LHA
                        INNER JOIN SI_autores SIA ON SIA.id = LHA.id_autor
                        INNER JOIN SI_editoriales SIE ON SIE.id = LHA.id_editorial
@@ -459,6 +544,7 @@ connection);
                     result.LengthHours = Convert.ToInt32(reader[4]);
                     result.LengthMins = Convert.ToInt32(reader[5]);
                     result.Title = reader[1].ToString().Trim();
+                    result.Date = DateTime.Parse(reader[8].ToString()).ToString("dd/MM/yyyy");
                 }
             }
 
@@ -571,20 +657,8 @@ ORDER BY Mm.f_alta DESC, id DESC",
             return result;
         }
 
-        public async Task<SuscriptionTitleLinkResult> GetSuscriptionTitleLink(string session, string id, int app = 1)
+        public async Task IncreaseSuscriptionTitleDownloadCounter(string session, string id, int app = 1)
         {
-            IAmazonS3 clientS3 = new AmazonS3Client(
-                new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey),
-                RegionEndpoint.EUWest1
-            );
-
-            var link = clientS3.GetPreSignedURL(new GetPreSignedUrlRequest
-            {
-                BucketName = _settings.AWSBucket,
-                Expires = DateTime.Now.AddSeconds(_settings.AWSLinkExpireInSecs),
-                Key = $"{id}.zip"
-            });
-
             // Update Database counter so we have to get user from dynamod db (from session)
             var currentUser = await GetUserFromSession(session);
 
@@ -603,6 +677,18 @@ ORDER BY Mm.f_alta DESC, id DESC",
                     await commandInsert.ExecuteNonQueryAsync();
                 }
             }
+        }
+
+        public SuscriptionTitleLinkResult GetSuscriptionTitleLink(string session, string id, int app = 1)
+        {
+            IAmazonS3 clientS3 = new AmazonS3Client();
+
+            var link = clientS3.GetPreSignedURL(new GetPreSignedUrlRequest
+            {
+                BucketName = _settings.AWSBucket,
+                Expires = DateTime.Now.AddSeconds(_settings.AWSLinkExpireInSecs),
+                Key = $"{id}.zip"
+            });
 
             return new SuscriptionTitleLinkResult
             {
@@ -625,14 +711,6 @@ connection);
 
             await commandInsert.ExecuteNonQueryAsync();
 
-            // Send SNS notification to Android and iOS registered devices
-            var credentials = new BasicAWSCredentials(_settings.AWSAccessKey, _settings.AWSSecretKey);
-
-            AmazonSimpleNotificationServiceClient snsClient = new AmazonSimpleNotificationServiceClient(
-                credentials,
-                RegionEndpoint.EUWest1
-            );
-
             // Get notification type based on id
             var code = "";
             if (id_suscripcion > 0)
@@ -646,7 +724,7 @@ connection);
                 topicArn += $"-{code}";
             }
 
-            var responseNotification = await snsClient.PublishAsync(new PublishRequest
+            var responseNotification = await _clientSNS.PublishAsync(new PublishRequest
             {
                 MessageStructure = "json",
                 TopicArn = topicArn,
@@ -660,7 +738,7 @@ connection);
 
             if (responseNotification.HttpStatusCode != System.Net.HttpStatusCode.OK)
             {
-                throw new InternalException(ExceptionCodes.PUBLISH_MESSAGE_ERROR, $"Error publishing extra notification message to Android: {responseNotification.HttpStatusCode}, {responseNotification.MessageId}");
+                throw new InternalException(ExceptionCodes.PUBLISH_MESSAGE_ERROR, $"Error publishing extra notification message: {responseNotification.HttpStatusCode}, {responseNotification.MessageId}");
             }
         }
 
